Clean up widgets, fix resize with objects on map
